Vertical stacking algorithm and borders
Added an algorithm to check when covers overlap and adjust their
vertical height so they don't anymore. It keeps a sorted list of cover
x positions, and when a new height at an x position is requested, it
finds that position in the list. Then it checks what heights are taken
on the left and right (because it's sorted, left and right may be
unnecessary, or maybe even undesirable! Things to ponder for the
future, perhaps.)

- Removed browser style padding/margins on `body`
- Increased the SVG size to be big, so more was onscreen.
- Increased the cover size to 60.
- Added a border around the covers.
- Put the covers and borders in a group element, and used
  transformation to put them in the correct location.