Require d3 and css through webpack
Instead of including a CDN version of d3, now it's bundled with the rest
of my code. The styles are also included now. Webpack is magic.

Switched the .gitignore from dist/bundle.js to just dist/ because of
sourcemaps. Since index.html was already in the repo it'll stay tracked.